Making mounts order available only for 8.7 and above (#440)
* mounts only 8.7 and up

+++ Lua/TFS_10/talkaction shopsystem/znoteshop.lua
@@ -76,16 +76,18 @@ function onSay(player, words, param)
 					end
 				end
 
-				-- ORDER TYPE 6 (Mounts)
-				if q_type == 6 then
-					served = true
-					-- Make sure player don't already have this outfit and addon
-					if not player:hasMount(q_itemid) then
-						db.query("DELETE FROM `znote_shop_orders` WHERE `id` = " .. q_id .. ";")
-						player:addMount(q_itemid)
-						player:sendTextMessage(MESSAGE_INFO_DESCR, "Congratulations! You have received a new mount!")
-					else
-						player:sendTextMessage(MESSAGE_STATUS_WARNING, "You already have this mount!")
+				if Game.getClientVersion().min >= 870 then
+					-- ORDER TYPE 6 (Mounts)
+					if q_type == 6 then
+						served = true
+						-- Make sure player don't already have this outfit and addon
+						if not player:hasMount(q_itemid) then
+							db.query("DELETE FROM `znote_shop_orders` WHERE `id` = " .. q_id .. ";")
+							player:addMount(q_itemid)
+							player:sendTextMessage(MESSAGE_INFO_DESCR, "Congratulations! You have received a new mount!")
+						else
+							player:sendTextMessage(MESSAGE_STATUS_WARNING, "You already have this mount!")
+						end
 					end
 				end
